+++ src/ducks/Studio/Sidebar/index.js
@@ -1,6 +1,7 @@
-import React, { useEffect } from 'react'
+import React, { useState, useEffect } from 'react'
 import cx from 'classnames'
 import Loader from '@santiment-network/ui/Loader/Loader'
+import Tabs from '@santiment-network/ui/Tabs'
 import Icon from '@santiment-network/ui/Icon'
 import MetricSelector from './MetricSelector'
 import Search from './Search'
@@ -8,29 +9,37 @@ import withMetrics from '../withMetrics'
 import { rebuildDescriptions } from '../../dataHub/metrics/descriptions'
 import styles from './index.module.scss'
 
-const Header = props => {
-  return (
-    <div className={styles.header}>
-      <h2 className={styles.title}>Metrics</h2>
-      <Search {...props} />
-    </div>
-  )
-}
+const TABS = ['Metrics', 'Insights']
+const DEFAULT_TAB = TABS[0]
 
-const CloseButton = ({ onClick, className }) => {
-  return (
-    <div className={cx(styles.toggle, className)} onClick={onClick}>
-      <div className={styles.close}>
-        <Icon type='sidebar' className={styles.icon} />
-      </div>
+const Header = ({ activeTab, setActiveTab, ...props }) => (
+  <div className={styles.header}>
+    <Tabs
+      options={TABS}
+      className={styles.tabs}
+      classes={styles}
+      defaultSelectedIndex={activeTab}
+      // NOTE: Not passed as a reference, since more than 1 argument is passed to a callback [@vanguard | Aug  4, 2020]
+      onSelect={tab => setActiveTab(tab)}
+    />
+    <Search {...props} />
+  </div>
+)
+
+const CloseButton = ({ onClick, className }) => (
+  <div className={cx(styles.toggle, className)} onClick={onClick}>
+    <div className={styles.close}>
+      <Icon type='sidebar' className={styles.icon} />
     </div>
-  )
-}
+  </div>
+)
 
 const Sidebar = ({ loading, children, ...rest }) => {
+  const [activeTab, setActiveTab] = useState(DEFAULT_TAB)
+
   return (
     <aside className={styles.wrapper}>
-      <Header {...rest} />
+      <Header activeTab={activeTab} setActiveTab={setActiveTab} {...rest} />
       <div className={styles.selector}>
         {loading ? (
           <Loader className={styles.loading} />
